Rewrite to use new API and look more like Python
This is embarrassingly simple with Translink's newly provided API,
so much of the code for crudely scraping HTML and JSON has been
ripped out.

Also changed a whole bunch of formatting and style to look
more like Python.

Parsing the information from the API's return data structure in
printStops.print_route_info() is the largest fragile part left.
To be improved.